requirements.txt
accelerate
codetiming
datasets
flash-attn
liger-kernel
mathruler
numpy
omegaconf
pandas
peft
pillow
pyarrow>=15.0.0
pylatexenc
qwen-vl-utils
ray
tensordict
transformers
vllm
wandb
sam2
matplotlib